Migration to the new layout
More info here https://www.youtube.com/watch?v=6mxWayq-s9I
