--- .config/nvim/after/plugin/colorscheme.lua
@@ -1,2 +1,0 @@
---vim.cmd [[colorscheme tokyonight]]
-vim.cmd [[colorscheme catppuccin]]
