--- .config/nvim/lua/main/copilot.lua
@@ -1,38 +1,0 @@
-local M = {}
-
-function M.setup()
-  require('copilot').setup({
-    panel = {
-      enabled = false,
-      auto_refresh = false,
-      keymap = {
-        jump_prev = "[[",
-        jump_next = "]]",
-        accept = "<CR>",
-        refresh = "gr",
-        open = "<M-CR>"
-      },
-    },
-    suggestion = {
-      -- ghost text suggestion turn off for cmp
-      enabled = true,
-      auto_trigger = true,
-      debounce = 75,
-      keymap = {
-        accept = "<Tab>",
-        accept_word = false,
-        accept_line = false,
-        next = "<M-]>",
-        prev = "<M-[>",
-        dismiss = "<C-]>",
-      },
-    },
-    filetypes = {
-      markdown = false,
-      help = false,
-    },
-  })
-  -- require('copilot_cmp').setup()
-end
-
-return M
