+++ .config/nvim/lua/main/core/init.lua
@@ -1,0 +1,3 @@
+require("main.core.remap")
+require("main.core.set")
+require("main.core.command")
