--- .config/nvim/lua/main/fugitive.lua
@@ -1,7 +1,0 @@
-local M = {}
-
-function M.setup()
-  vim.keymap.set("n", "<leader>vs", ":tab Git<CR>")
-end
-
-return M
