+++ .config/nvim/lua/main/init.lua
@@ -1,3 +1,2 @@
-require("main.set")
-require("main.remap")
-require("main.command")
+require("main.core")
+require("main.lazy")
