+++ .config/nvim/lua/main/lazy.lua
@@ -1,0 +1,24 @@
+local lazypath = vim.fn.stdpath("data") .. "/lazy/lazy.nvim"
+if not (vim.uv or vim.loop).fs_stat(lazypath) then
+  vim.fn.system({
+    "git",
+    "clone",
+    "--filter=blob:none",
+    "https://github.com/folke/lazy.nvim.git",
+    "--branch=stable", -- latest stable release
+    lazypath,
+  })
+end
+
+vim.opt.rtp:prepend(lazypath)
+
+local setup = {
+  { import = "main.plugins" }
+}
+
+local opts = {
+  ui = { border = "rounded" }
+}
+
+require("lazy").setup(setup, opts)
+
