--- .config/nvim/lua/main/neotree.lua
@@ -1,29 +1,0 @@
-
-local M = {}
-
-function M.setup()
-
-  local builtin = require('neo-tree.command')
-  vim.keymap.set("n", "\\", function() builtin.execute({ toggle = true, reveal = true }) end )
-
-  require("neo-tree").setup({
-    close_if_last_window = true,
-    filesystem = {
-      filtered_items = {
-        visible = true,
-        hide_dotfiles = false,
-        hide_gitignored = true,
-      },
-    },
-    event_handlers = {
-      {
-        event = "file_opened",
-        handler = function(_)
-         builtin.execute({ action = "close" })
-        end
-      },
-    },
-  })
-end
-
-return M
