+++ .config/nvim/lua/main/plugins/autopairs.lua
@@ -1,0 +1,12 @@
+return {
+  'windwp/nvim-autopairs',
+  module = {'nvim-autopairs.completion.cmp', 'nvim-autopairs'},
+  config = function ()
+    local npairs = require('nvim-autopairs')
+    npairs.setup {
+      check_ts = true,
+    }
+    npairs.add_rules(require 'nvim-autopairs.rules.endwise-lua')
+  end
+}
+
