+++ .config/nvim/lua/main/plugins/colorscheme.lua
@@ -1,0 +1,7 @@
+return {
+  'catppuccin/nvim',
+  name = 'catppuccin',
+  config = function()
+    vim.cmd([[colorscheme catppuccin]])
+  end,
+}
