+++ .config/nvim/lua/main/plugins/copilot-chat.lua
@@ -1,9 +1,8 @@
-local M = {}
-
-function M.setup()
-  local copilot_chat = require("CopilotChat")
-
-  copilot_chat.setup({
+-- https://github.com/CopilotC-Nvim/CopilotChat.nvim
+return {
+  'CopilotC-Nvim/CopilotChat.nvim',
+  branch = 'canary',
+  opts = {
     context = 'buffers',
     mappings = {
       complete = {
@@ -39,39 +38,42 @@ function M.setup()
         normal = 'gs'
       }
     }
-  })
+  },
+  config = function()
+    local copilot_chat = require("CopilotChat")
 
-  -- Toggle the chat window
-  vim.keymap.set("n", "<Leader>cct", function()
-        copilot_chat.toggle()
-  end, {})
+    -- Toggle the chat window
+    vim.keymap.set("n", "<Leader>cct", function()
+      copilot_chat.toggle()
+    end, {})
 
-  -- Ask a question for the current buffer
-  vim.keymap.set("n", "<Leader>ccq", function()
+    -- Ask a question for the current buffer
+    vim.keymap.set("n", "<Leader>ccq", function()
       local input = vim.fn.input("Quick Chat: ")
       if input ~= "" then
         copilot_chat.ask(input, { selection = require("CopilotChat.select").buffer })
       end
-  end, {})
+    end, {})
 
-  -- Ask a question for selection in the current buffer
-  vim.keymap.set("v", "<Leader>ccq", function()
+    -- Ask a question for selection in the current buffer
+    vim.keymap.set("v", "<Leader>ccq", function()
       local input = vim.fn.input("Quick Chat: ")
       if input ~= "" then
         copilot_chat.ask(input, { selection = require("CopilotChat.select").visual })
       end
-  end, {})
+    end, {})
 
-  vim.keymap.set("n", "<Leader>cch", function()
+    vim.keymap.set("n", "<Leader>cch", function()
       local actions = require("CopilotChat.actions")
       copilot_chat.pick(actions.help_actions())
-  end, {})
+    end, {})
 
-  vim.keymap.set("n", "<Leader>ccp", function()
+    vim.keymap.set("n", "<Leader>ccp", function()
       local actions = require("CopilotChat.actions")
       copilot_chat.pick(actions.prompt_actions())
-  end, {})
+    end, {})
 
-end
+  end,
+}
 
-return M
+
