+++ .config/nvim/lua/main/plugins/fugitive.lua
@@ -1,0 +1,6 @@
+return {
+  'tpope/vim-fugitive',
+  config = function()
+    vim.keymap.set("n", "<leader>vs", ":tab Git<CR>")
+  end,
+}
