+++ .config/nvim/lua/main/plugins/fzf.lua
@@ -1,0 +1,5 @@
+return {
+  'junegunn/fzf',
+  build = './install --all',
+  enabled = true,
+}
