+++ .config/nvim/lua/main/plugins/lualine.lua
@@ -1,0 +1,15 @@
+-- https://github.com/nvim-lualine/lualine.nvim
+return {
+  'nvim-lualine/lualine.nvim',
+  event = 'VimEnter',
+  config = function()
+    require('lualine').setup {
+      options = {
+--        theme = 'tokyonight',
+        section_separators = '',
+        component_separators = ''
+      }
+    }
+  end,
+}
+
