+++ .config/nvim/lua/main/plugins/undotree.lua
@@ -1,0 +1,7 @@
+-- https://github.com/mbbill/undotree
+return {
+  "mbbill/undotree",
+  config = function()
+    vim.keymap.set("n", "<leader>u", ":UndotreeToggle<CR>:UndotreeFocus<CR>")
+  end,
+}
