--- .config/nvim/lua/main/undotree.lua
@@ -1,7 +1,0 @@
-local M = {}
-
-function M.setup()
-  vim.keymap.set("n", "<leader>u", ":UndotreeToggle<CR>:UndotreeFocus<CR>")
-end
-
-return M
